test/requirements.txt
umap-learn~=0.5.0
numpy~=1.26.0
